LayoutTests/js/set-keys-expected.txt

Tests basic correctness of ES Set's keys() API

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS Set.prototype.keys.length is 0
PASS Set.prototype.keys === Set.prototype.values is true
PASS successfullyParsed is true

TEST COMPLETE
